Tests Project completely. inversify.config exports ContainerBuilder, so that multiple users will create independent containers. Bugfixes call to ts.createProgram (file paths were project-relative, which failed when cwd !== projectRoot. Works now with absolute paths).

+++ build/spec/helpers/copy-mock-project.js
@@ -4,4 +4,6 @@ var path = require("path");
 /**
  * At runtime, __dirname === @root/build/spec/helpers
  */
-fs.copySync(path.join(__dirname, '../../../spec/mock-project'), path.join(__dirname, '../mock-project'));
+if (!fs.existsSync(path.join(__dirname, '../mock-project'))) {
+    fs.copySync(path.join(__dirname, '../../../spec/mock-project'), path.join(__dirname, '../mock-project'));
+}
